AoC 2023 Day 4: Prepare sample input for part 2
Also give part 1 input new name. Why is it that on the first day
they actually introduce a different sample input for part 2, I
decide that "surely there is a pattern now that there is only one
sample input, so let me just name it sample-input.txt to keep it
simple"?